Added a stringify function to avoid unneeded recursion when scanning code. (#12858)
When an invocation is complex enough, it will go further than our logic in the analyzer, but we don't need to follow the properties or class names based on other variables - we still won't find out whether the class or property was moved in Jetpack without runtime analysis. Therefore we replace these instances with just a node's class name for want of a better string.